requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Video processing
av>=10.0.0
decord>=0.6.0

# Grad-CAM and visualization
grad-cam>=1.4.8
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
pandas>=2.0.0
scikit-learn>=1.3.0

# Training utilities
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Jupyter for notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
